scripting/perl: Forward compatibility for hooks
Currently the perl script to generate the hooks for the perl API must be
updated whenever a new hook is added, either to support the hook or to
explicitly unsupport it, as doing neither results in a compilation error
of the generated file perl_hooks.c.

This patch updates the script to prevent hooks from being added to the
hooks list if the argument type is unsupported and emits a warning if
the hook has not been explicitly marked as unsupported.

(cherry-picked from commit 015855ef4b843764cd8a0fac3153ce1b662b4813)